+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added `setESType`, `copyTo`, and `isStringlike` helper functions for schema conversion.
 - The following fields are automatically included for `passthrough`: `type`, `fields`, `copy_to`.
+- Added `nested` flag to `convertSchema` to support the nested type.
 
 # 0.54.0